Cleaned up Client Class
Client Class now has a global (static) Socket variable, which all
methods are capable of accessing.
StartClient now takes an IP (string) that is parsed for the connection.

Haven't tested to see if it actually connects to server, but it should.
If there's a problem, contact me, I know what could be wrong (if it is).